restructure folders name
